requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.3
django-cors-headers==3.14.0
django-environ==0.10.0
django-livesync==0.5
djangorestframework==3.13.1
djangorestframework-camel-case==1.4.2
future==0.18.3
mysqlclient==2.1.1
Pillow==9.4.0
pytz==2022.1
serializers==0.2.4
sqlparse==0.4.2
tornado==6.2
watchdog==2.2.0
websocket-client==1.4.2
whitenoise==6.4.0
